FEATURE/MEDIUM: check_disk_smart: Add mode to get data via SNMP extend OID
Tagged as MEDIUM because SSH long options --user and --identity were
renamed respectively --ssh-user and --ssh-identity.

When using option -O, --snmp-oid, common option are ignored and the
result is retrieved from the given OID. The result is expected to be in
get_smart format, one device per line: <device>\t<status>, eg:
/dev/sda	PASSED
/dev/sdb	PASSED

A sample script for the snmpd extend OID is printed on stdout with option
-T, --snmp-sample-script. Put it on the target host and add an extend line
in your snmpd.conf to access its output.